refactor(api): extract compile task pipeline from API
- move compile task hooks and task state into the api.task package
- move compile pipeline execution and diagnostic remapping into CompileTaskRunner
- keep API focused on facade orchestration and update API test support imports
