+++ src/main/java/dev/superice/gdcc/api/API.java
@@ -1,34 +1,21 @@
 package dev.superice.gdcc.api;
 
-import dev.superice.gdcc.backend.CodegenContext;
+import dev.superice.gdcc.api.task.CompileTaskHooks;
+import dev.superice.gdcc.api.task.CompileTaskRunner;
+import dev.superice.gdcc.api.task.CompileTaskState;
 import dev.superice.gdcc.backend.c.build.CProjectBuilder;
-import dev.superice.gdcc.backend.c.build.CProjectInfo;
-import dev.superice.gdcc.backend.c.gen.CCodegen;
 import dev.superice.gdcc.exception.ApiCompileAlreadyRunningException;
 import dev.superice.gdcc.exception.ApiCompileTaskNotFoundException;
-import dev.superice.gdcc.exception.ApiEntryTypeMismatchException;
 import dev.superice.gdcc.exception.ApiModuleAlreadyExistsException;
 import dev.superice.gdcc.exception.ApiModuleBusyException;
 import dev.superice.gdcc.exception.ApiModuleNotFoundException;
-import dev.superice.gdcc.frontend.diagnostic.DiagnosticManager;
-import dev.superice.gdcc.frontend.diagnostic.FrontendDiagnostic;
-import dev.superice.gdcc.frontend.diagnostic.DiagnosticSnapshot;
 import dev.superice.gdcc.frontend.lowering.FrontendLoweringPassManager;
-import dev.superice.gdcc.frontend.parse.FrontendModule;
-import dev.superice.gdcc.frontend.parse.FrontendSourceUnit;
 import dev.superice.gdcc.frontend.parse.GdScriptParserService;
-import dev.superice.gdcc.gdextension.ExtensionApiLoader;
-import dev.superice.gdcc.scope.ClassRegistry;
 import dev.superice.gdcc.util.StringUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.time.Clock;
-import java.time.Instant;
-import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
@@ -36,7 +23,6 @@ import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Function;
-import java.util.function.Consumer;
 
 /// In-memory module registry facade intended for RPC adapters.
 ///
@@ -44,14 +30,6 @@ import java.util.function.Consumer;
 /// backend remain the sole compilation fact sources. Step 2 adds virtual-path normalization plus
 /// in-memory file/directory CRUD without coupling the facade to any transport framework.
 public final class API {
-    private static final @NotNull DiagnosticSnapshot EMPTY_DIAGNOSTICS = new DiagnosticSnapshot(List.of());
-    private static final @NotNull List<String> GENERATED_FILE_NAMES = List.of(
-            "entry.c",
-            "engine_method_binds.h",
-            "entry.h"
-    );
-    private static final @NotNull ThreadLocal<CompileTaskState> CURRENT_COMPILE_TASK = new ThreadLocal<>();
-
     private final @NotNull Clock clock;
     private final @NotNull GdScriptParserService parserService;
     private final @NotNull FrontendLoweringPassManager loweringPassManager;
@@ -61,33 +39,6 @@ public final class API {
     private final @NotNull ConcurrentHashMap<Long, CompileTaskState> compileTasks = new ConcurrentHashMap<>();
     private final @NotNull ConcurrentHashMap<String, Long> activeCompileTasksByModule = new ConcurrentHashMap<>();
     private final @NotNull AtomicLong nextCompileTaskId = new AtomicLong(1);
-
-    static final class CompileTaskHooks {
-        private static final @NotNull CompileTaskHooks NONE = new CompileTaskHooks(null);
-
-        private final @Nullable Consumer<CompileTaskSnapshot> afterSnapshotUpdate;
-
-        private CompileTaskHooks(@Nullable Consumer<CompileTaskSnapshot> afterSnapshotUpdate) {
-            this.afterSnapshotUpdate = afterSnapshotUpdate;
-        }
-
-        static @NotNull CompileTaskHooks none() {
-            return NONE;
-        }
-
-        static @NotNull CompileTaskHooks afterSnapshotUpdate(@NotNull Consumer<CompileTaskSnapshot> afterSnapshotUpdate) {
-            return new CompileTaskHooks(Objects.requireNonNull(
-                    afterSnapshotUpdate,
-                    "afterSnapshotUpdate must not be null"
-            ));
-        }
-
-        void afterSnapshotUpdate(@NotNull CompileTaskSnapshot snapshot) {
-            if (afterSnapshotUpdate != null) {
-                afterSnapshotUpdate.accept(snapshot);
-            }
-        }
-    }
 
     public API() {
         this(
@@ -126,13 +77,7 @@ public final class API {
     /// Records one event for the compile task currently executing on this thread. This is meant for
     /// frontend/backend code that runs inside the API-managed compile pipeline.
     public static boolean recordCurrentCompileTaskEvent(@NotNull String category, @NotNull String detail) {
-        var event = new CompileTaskEvent(category, detail);
-        var taskState = CURRENT_COMPILE_TASK.get();
-        if (taskState == null) {
-            return false;
-        }
-        taskState.recordEvent(event);
-        return true;
+        return CompileTaskState.recordCurrentThreadEvent(category, detail);
     }
 
     public @NotNull ModuleSnapshot createModule(@NotNull String moduleId, @NotNull String moduleName) {
@@ -318,12 +263,25 @@ public final class API {
             activeCompileTasksByModule.remove(normalizedModuleId, taskId);
             throw exception;
         }
+        var ownerState = managedModule.state();
         var taskState = new CompileTaskState(taskId, normalizedModuleId, clock.instant(), compileTaskHooks);
         compileTasks.put(taskId, taskState);
         try {
             Thread.ofVirtual()
                     .name("gdcc-api-compile-" + taskId)
-                    .start(() -> runCompileTask(taskState, normalizedModuleId, managedModule));
+                    .start(new CompileTaskRunner(
+                            clock,
+                            parserService,
+                            loweringPassManager,
+                            projectBuilder,
+                            taskState,
+                            () -> freezeCompileTaskRequest(ownerState),
+                            result -> {
+                                ownerState.setLastCompileResult(result);
+                                activeCompileTasksByModule.remove(normalizedModuleId, taskId);
+                                managedModule.endCompile(taskId);
+                            }
+                    ));
         } catch (RuntimeException exception) {
             compileTasks.remove(taskId);
             activeCompileTasksByModule.remove(normalizedModuleId, taskId);
@@ -375,419 +333,30 @@ public final class API {
         return taskState;
     }
 
-    /// Compilation runs against a frozen module snapshot so RPC writes after `compile(...)` starts
-    /// cannot mutate the exact source set or compile options consumed by this pipeline.
-    private @NotNull CompileResult compileFrozenRequest(
-            @NotNull ModuleState ownerState,
-            @NotNull ModuleState.CompileRequest request,
-            @NotNull CompileTaskState taskState
-    ) {
-        var sourcePaths = displaySourcePaths(request);
-        try {
-            ownerState.prepareOutputMountRoot(request.compileOptions().outputMountRoot());
-        } catch (ApiEntryTypeMismatchException | IllegalArgumentException exception) {
-            return new CompileResult(
-                    CompileResult.Outcome.CONFIGURATION_FAILED,
-                    request.compileOptions(),
-                    request.topLevelCanonicalNameMap(),
-                    sourcePaths,
-                    EMPTY_DIAGNOSTICS,
-                    "Compile outputs for module '"
-                            + request.moduleId()
-                            + "' could not be mounted: "
-                            + exception.getMessage(),
-                    "",
-                    List.of(),
-                    List.of(),
-                    List.of()
-            );
-        }
-        if (request.failure() != null) {
-            return new CompileResult(
-                    request.failure().outcome(),
-                    request.compileOptions(),
-                    request.topLevelCanonicalNameMap(),
-                    sourcePaths,
-                    EMPTY_DIAGNOSTICS,
-                    request.failure().message(),
-                    "",
-                    List.of(),
-                    List.of(),
-                    List.of()
-            );
-        }
-        if (request.sourceSnapshots().isEmpty()) {
-            return new CompileResult(
-                    CompileResult.Outcome.SOURCE_COLLECTION_FAILED,
-                    request.compileOptions(),
-                    request.topLevelCanonicalNameMap(),
-                    sourcePaths,
-                    EMPTY_DIAGNOSTICS,
-                    "Module '" + request.moduleId() + "' has no .gd source files to compile",
-                    "",
-                    List.of(),
-                    List.of(),
-                    List.of()
-            );
-        }
-
-        var projectPath = request.compileOptions().projectPath();
-        if (projectPath == null) {
-            return new CompileResult(
-                    CompileResult.Outcome.CONFIGURATION_FAILED,
-                    request.compileOptions(),
-                    request.topLevelCanonicalNameMap(),
-                    sourcePaths,
-                    EMPTY_DIAGNOSTICS,
-                    "Compile options for module '" + request.moduleId() + "' must set projectPath before compile",
-                    "",
-                    List.of(),
-                    List.of(),
-                    List.of()
-            );
-        }
-
-        var diagnostics = new DiagnosticManager();
-        taskState.updateRunningStage(
-                CompileTaskSnapshot.Stage.PARSING,
-                "Parsing source units",
-                0,
-                request.sourceSnapshots().size(),
-                null
-        );
-        var units = new ArrayList<FrontendSourceUnit>(request.sourceSnapshots().size());
-        for (var index = 0; index < request.sourceSnapshots().size(); index++) {
-            var sourceSnapshot = request.sourceSnapshots().get(index);
-            taskState.updateRunningStage(
-                    CompileTaskSnapshot.Stage.PARSING,
-                    "Parsing " + sourceSnapshot.displayPath(),
-                    index,
-                    request.sourceSnapshots().size(),
-                    sourceSnapshot.displayPath()
-            );
-            units.add(parserService.parseUnit(
-                    sourceSnapshot.logicalPath(),
-                    sourceSnapshot.source(),
-                    diagnostics
-            ));
-            taskState.updateRunningStage(
-                    CompileTaskSnapshot.Stage.PARSING,
-                    "Parsed " + sourceSnapshot.displayPath(),
-                    index + 1,
-                    request.sourceSnapshots().size(),
-                    sourceSnapshot.displayPath()
-            );
-        }
-        var parseAndFrontendDiagnostics = remapDiagnosticSourcePaths(request, diagnostics.snapshot());
-        if (parseAndFrontendDiagnostics.hasErrors()) {
-            return new CompileResult(
-                    CompileResult.Outcome.FRONTEND_FAILED,
-                    request.compileOptions(),
-                    request.topLevelCanonicalNameMap(),
-                    sourcePaths,
-                    parseAndFrontendDiagnostics,
-                    "Frontend diagnostics blocked compilation",
-                    "",
-                    List.of(),
-                    List.of(),
-                    List.of()
-            );
-        }
-
-        var projectPathError = ensureProjectPathUsable(projectPath);
-        if (projectPathError != null) {
-            return new CompileResult(
-                    CompileResult.Outcome.CONFIGURATION_FAILED,
-                    request.compileOptions(),
-                    request.topLevelCanonicalNameMap(),
-                    sourcePaths,
-                    parseAndFrontendDiagnostics,
-                    projectPathError,
-                    "",
-                    List.of(),
-                    List.of(),
-                    List.of()
-            );
-        }
-
-        try {
-            var extensionApi = ExtensionApiLoader.loadVersion(request.compileOptions().godotVersion());
-            var classRegistry = new ClassRegistry(extensionApi);
-            var frontendModule = new FrontendModule(
-                    request.moduleName(),
-                    units,
-                    request.topLevelCanonicalNameMap()
-            );
-            taskState.updateRunningStage(
-                    CompileTaskSnapshot.Stage.LOWERING,
-                    "Lowering frontend module",
-                    request.sourceSnapshots().size(),
-                    request.sourceSnapshots().size(),
-                    null
-            );
-            var lowered = loweringPassManager.lower(frontendModule, classRegistry, diagnostics);
-            var frontendDiagnostics = remapDiagnosticSourcePaths(request, diagnostics.snapshot());
-            if (lowered == null || frontendDiagnostics.hasErrors()) {
-                return new CompileResult(
-                        CompileResult.Outcome.FRONTEND_FAILED,
-                        request.compileOptions(),
-                        request.topLevelCanonicalNameMap(),
-                        sourcePaths,
-                        frontendDiagnostics,
-                        "Frontend diagnostics blocked compilation",
-                        "",
-                        List.of(),
-                        List.of(),
-                        List.of()
-                );
-            }
-
-            var projectInfo = new CProjectInfo(
-                    request.moduleId(),
-                    request.compileOptions().godotVersion(),
-                    projectPath,
-                    request.compileOptions().optimizationLevel(),
-                    request.compileOptions().targetPlatform()
-            );
-            var codegen = new CCodegen();
-            taskState.updateRunningStage(
-                    CompileTaskSnapshot.Stage.CODEGEN_PREPARE,
-                    "Preparing C code generation",
-                    request.sourceSnapshots().size(),
-                    request.sourceSnapshots().size(),
-                    null
-            );
-            codegen.prepare(
-                    new CodegenContext(projectInfo, classRegistry, request.compileOptions().strictMode()),
-                    lowered
-            );
-
-            taskState.updateRunningStage(
-                    CompileTaskSnapshot.Stage.BUILDING_NATIVE,
-                    "Building native artifacts",
-                    request.sourceSnapshots().size(),
-                    request.sourceSnapshots().size(),
-                    null
-            );
-            var buildResult = projectBuilder.buildProject(projectInfo, codegen);
-            var generatedFiles = collectGeneratedFiles(projectPath);
-            if (!buildResult.success()) {
-                return new CompileResult(
-                        CompileResult.Outcome.BUILD_FAILED,
-                        request.compileOptions(),
-                        request.topLevelCanonicalNameMap(),
-                        sourcePaths,
-                        frontendDiagnostics,
-                        "Native build reported failure; see buildLog for details",
-                        buildResult.buildLog(),
-                        generatedFiles,
-                        buildResult.artifacts(),
-                        List.of()
-                );
-            }
-            try {
-                var outputLinks = ownerState.mountCompileOutputs(
+    private @NotNull CompileTaskRunner.Request freezeCompileTaskRequest(@NotNull ModuleState ownerState) {
+        var request = ownerState.freezeCompileRequest();
+        return new CompileTaskRunner.Request(
+                request.moduleId(),
+                request.moduleName(),
+                request.compileOptions(),
+                request.topLevelCanonicalNameMap(),
+                request.sourceSnapshots().stream()
+                        .map(sourceSnapshot -> new CompileTaskRunner.SourceSnapshot(
+                                sourceSnapshot.displayPath(),
+                                sourceSnapshot.logicalPath(),
+                                sourceSnapshot.source()
+                        ))
+                        .toList(),
+                request.failure() == null
+                        ? null
+                        : new CompileTaskRunner.Failure(request.failure().outcome(), request.failure().message()),
+                ownerState::prepareOutputMountRoot,
+                outputs -> ownerState.mountCompileOutputs(
                         request.compileOptions().outputMountRoot(),
-                        generatedFiles,
-                        buildResult.artifacts()
-                );
-                return new CompileResult(
-                        CompileResult.Outcome.SUCCESS,
-                        request.compileOptions(),
-                        request.topLevelCanonicalNameMap(),
-                        sourcePaths,
-                        frontendDiagnostics,
-                        null,
-                        buildResult.buildLog(),
-                        generatedFiles,
-                        buildResult.artifacts(),
-                        outputLinks
-                );
-            } catch (ApiEntryTypeMismatchException | IllegalArgumentException exception) {
-                return new CompileResult(
-                        CompileResult.Outcome.CONFIGURATION_FAILED,
-                        request.compileOptions(),
-                        request.topLevelCanonicalNameMap(),
-                        sourcePaths,
-                        frontendDiagnostics,
-                        "Compile outputs for module '"
-                                + request.moduleId()
-                                + "' could not be mounted: "
-                                + exception.getMessage(),
-                        buildResult.buildLog(),
-                        generatedFiles,
-                        buildResult.artifacts(),
-                        List.of()
-                );
-            }
-        } catch (IOException exception) {
-            return new CompileResult(
-                    CompileResult.Outcome.BUILD_FAILED,
-                    request.compileOptions(),
-                    request.topLevelCanonicalNameMap(),
-                    sourcePaths,
-                    remapDiagnosticSourcePaths(request, diagnostics.snapshot()),
-                    "Build pipeline failed: " + exception.getMessage(),
-                    exception.getMessage(),
-                    collectGeneratedFiles(projectPath),
-                    List.of(),
-                    List.of()
-            );
-        }
-    }
-
-    private @Nullable String ensureProjectPathUsable(@NotNull Path projectPath) {
-        try {
-            Files.createDirectories(projectPath);
-            return null;
-        } catch (IOException exception) {
-            return "Project path '" + projectPath + "' cannot be used as a build directory: " + exception.getMessage();
-        }
-    }
-
-    private @NotNull List<String> displaySourcePaths(@NotNull ModuleState.CompileRequest request) {
-        return request.sourceSnapshots().stream()
-                .map(ModuleState.SourceSnapshot::displayPath)
-                .toList();
-    }
-
-    /// Frontend internals still track host-usable logical paths, but API callers care about the
-    /// original caller-facing labels such as `res://player.gd`. Results therefore remap matching
-    /// frontend diagnostic source paths back to each frozen source snapshot's `displayPath`.
-    private @NotNull DiagnosticSnapshot remapDiagnosticSourcePaths(
-            @NotNull ModuleState.CompileRequest request,
-            @NotNull DiagnosticSnapshot diagnostics
-    ) {
-        if (diagnostics.isEmpty()) {
-            return diagnostics;
-        }
-        var displayPathsByLogicalPath = request.sourceSnapshots().stream()
-                .collect(java.util.stream.Collectors.toMap(
-                        sourceSnapshot -> FrontendDiagnostic.sourcePathText(sourceSnapshot.logicalPath()),
-                        ModuleState.SourceSnapshot::displayPath,
-                        (first, _) -> first
-                ));
-        var remappedDiagnostics = diagnostics.asList().stream()
-                .map(diagnostic -> remapDiagnosticSourcePath(diagnostic, displayPathsByLogicalPath))
-                .toList();
-        return new DiagnosticSnapshot(remappedDiagnostics);
-    }
-
-    private @NotNull FrontendDiagnostic remapDiagnosticSourcePath(
-            @NotNull FrontendDiagnostic diagnostic,
-            @NotNull Map<String, String> displayPathsByLogicalPath
-    ) {
-        var sourcePath = diagnostic.sourcePath();
-        if (sourcePath == null) {
-            return diagnostic;
-        }
-        var displayPath = displayPathsByLogicalPath.get(sourcePath);
-        if (Objects.equals(sourcePath, displayPath) || displayPath == null) {
-            return diagnostic;
-        }
-        return new FrontendDiagnostic(
-                diagnostic.severity(),
-                diagnostic.category(),
-                diagnostic.message(),
-                displayPath,
-                diagnostic.range()
+                        outputs.generatedFiles(),
+                        outputs.artifacts()
+                )
         );
-    }
-
-    private void runCompileTask(
-            @NotNull CompileTaskState taskState,
-            @NotNull String normalizedModuleId,
-            @NotNull ManagedModule ownerModule
-    ) {
-        ModuleState.CompileRequest request = null;
-        CURRENT_COMPILE_TASK.set(taskState);
-        try {
-            try {
-                taskState.updateRunningStage(
-                        CompileTaskSnapshot.Stage.FREEZING_INPUTS,
-                        "Freezing compile inputs",
-                        0,
-                        0,
-                        null
-                );
-                taskState.updateRunningStage(
-                        CompileTaskSnapshot.Stage.COLLECTING_SOURCES,
-                        "Collecting .gd sources from module VFS",
-                        0,
-                        0,
-                        null
-                );
-                request = ownerModule.state().freezeCompileRequest();
-                taskState.updateRunningStage(
-                        CompileTaskSnapshot.Stage.COLLECTING_SOURCES,
-                        "Collected " + request.sourceSnapshots().size() + " source units",
-                        0,
-                        request.sourceSnapshots().size(),
-                        null
-                );
-                var result = compileFrozenRequest(ownerModule.state(), request, taskState);
-                finishCompileTask(taskState, normalizedModuleId, ownerModule, result);
-            } catch (Throwable throwable) {
-                finishCompileTask(
-                        taskState,
-                        normalizedModuleId,
-                        ownerModule,
-                        unexpectedTaskFailure(request, throwable)
-                );
-            }
-        } finally {
-            CURRENT_COMPILE_TASK.remove();
-        }
-    }
-
-    private void finishCompileTask(
-            @NotNull CompileTaskState taskState,
-            @NotNull String normalizedModuleId,
-            @NotNull ManagedModule ownerModule,
-            @NotNull CompileResult result
-    ) {
-        taskState.complete(
-                clock.instant(),
-                result,
-                result.success() ? CompileTaskSnapshot.Stage.FINISHED : taskState.snapshot().stage()
-        );
-        ownerModule.state().setLastCompileResult(result);
-        activeCompileTasksByModule.remove(normalizedModuleId, taskState.taskId());
-        ownerModule.endCompile(taskState.taskId());
-    }
-
-    private @NotNull CompileResult unexpectedTaskFailure(
-            @Nullable ModuleState.CompileRequest request,
-            @NotNull Throwable throwable
-    ) {
-        var projectPath = request == null ? null : request.compileOptions().projectPath();
-        var failureMessage = "Compile task failed unexpectedly: " + describeThrowable(throwable);
-        return new CompileResult(
-                CompileResult.Outcome.BUILD_FAILED,
-                request == null ? CompileOptions.defaults() : request.compileOptions(),
-                request == null ? Map.of() : request.topLevelCanonicalNameMap(),
-                request == null ? List.of() : displaySourcePaths(request),
-                EMPTY_DIAGNOSTICS,
-                failureMessage,
-                failureMessage,
-                projectPath == null ? List.of() : collectGeneratedFiles(projectPath),
-                List.of(),
-                List.of()
-        );
-    }
-
-    private @NotNull String describeThrowable(@NotNull Throwable throwable) {
-        var message = throwable.getMessage();
-        return message == null || message.isBlank()
-                ? throwable.getClass().getSimpleName()
-                : throwable.getClass().getSimpleName() + ": " + message;
-    }
-
-    private @NotNull List<Path> collectGeneratedFiles(@NotNull Path projectPath) {
-        return GENERATED_FILE_NAMES.stream()
-                .map(projectPath::resolve)
-                .filter(Files::exists)
-                .toList();
     }
 
     /// First implementation favors correctness over read/write concurrency: all same-module API
@@ -921,147 +490,4 @@ public final class API {
         }
     }
 
-    private static final class CompileTaskState {
-        private final long taskId;
-        private final @NotNull String moduleId;
-        private final @NotNull Instant createdAt;
-        private final @NotNull CompileTaskHooks compileTaskHooks;
-        private final @NotNull ArrayList<CompileTaskEvent> events = new ArrayList<>();
-        private volatile @NotNull CompileTaskSnapshot snapshot;
-
-        private CompileTaskState(
-                long taskId,
-                @NotNull String moduleId,
-                @NotNull Instant createdAt,
-                @NotNull CompileTaskHooks compileTaskHooks
-        ) {
-            this.taskId = taskId;
-            this.moduleId = moduleId;
-            this.createdAt = createdAt;
-            this.compileTaskHooks = compileTaskHooks;
-            snapshot = new CompileTaskSnapshot(
-                    taskId,
-                    moduleId,
-                    CompileTaskSnapshot.State.QUEUED,
-                    CompileTaskSnapshot.Stage.QUEUED,
-                    "Queued for execution",
-                    0,
-                    0,
-                    null,
-                    1,
-                    createdAt,
-                    null,
-                    null
-            );
-        }
-
-        private long taskId() {
-            return taskId;
-        }
-
-        private @NotNull CompileTaskSnapshot snapshot() {
-            return snapshot;
-        }
-
-        private synchronized void recordEvent(@NotNull CompileTaskEvent event) {
-            events.add(Objects.requireNonNull(event, "event must not be null"));
-        }
-
-        private synchronized @Nullable CompileTaskEvent latestEvent() {
-            return events.isEmpty() ? null : events.getLast();
-        }
-
-        private synchronized @NotNull List<CompileTaskEvent> events() {
-            return List.copyOf(events);
-        }
-
-        private synchronized void clearEvents() {
-            events.clear();
-        }
-
-        private void updateRunningStage(
-                @NotNull CompileTaskSnapshot.Stage stage,
-                @Nullable String stageMessage,
-                int completedUnits,
-                int totalUnits,
-                @Nullable String currentSourcePath
-        ) {
-            var next = nextRunningSnapshot(stage, stageMessage, completedUnits, totalUnits, currentSourcePath);
-            if (next == null) {
-                return;
-            }
-            compileTaskHooks.afterSnapshotUpdate(next);
-        }
-
-        private synchronized @Nullable CompileTaskSnapshot nextRunningSnapshot(
-                @NotNull CompileTaskSnapshot.Stage stage,
-                @Nullable String stageMessage,
-                int completedUnits,
-                int totalUnits,
-                @Nullable String currentSourcePath
-        ) {
-            var current = snapshot;
-            if (current.completed()) {
-                throw new IllegalStateException("Completed compile task cannot return to RUNNING state");
-            }
-            var next = new CompileTaskSnapshot(
-                    taskId,
-                    moduleId,
-                    CompileTaskSnapshot.State.RUNNING,
-                    stage,
-                    stageMessage,
-                    completedUnits,
-                    totalUnits,
-                    currentSourcePath,
-                    current.revision() + 1,
-                    createdAt,
-                    null,
-                    null
-            );
-            if (sameProgress(current, next)) {
-                return null;
-            }
-            snapshot = next;
-            return next;
-        }
-
-        private void complete(
-                @NotNull Instant completedAt,
-                @NotNull CompileResult result,
-                @NotNull CompileTaskSnapshot.Stage completedStage
-        ) {
-            CompileTaskSnapshot next;
-            synchronized (this) {
-                var current = snapshot;
-                next = new CompileTaskSnapshot(
-                        taskId,
-                        moduleId,
-                        result.success() ? CompileTaskSnapshot.State.SUCCEEDED : CompileTaskSnapshot.State.FAILED,
-                        completedStage,
-                        result.success() ? "Compile completed successfully" : current.stageMessage(),
-                        result.success() ? current.totalUnits() : current.completedUnits(),
-                        current.totalUnits(),
-                        result.success() ? null : current.currentSourcePath(),
-                        current.revision() + 1,
-                        createdAt,
-                        completedAt,
-                        result
-                );
-                snapshot = next;
-            }
-            compileTaskHooks.afterSnapshotUpdate(next);
-        }
-
-        private boolean sameProgress(
-                @NotNull CompileTaskSnapshot current,
-                @NotNull CompileTaskSnapshot next
-        ) {
-            return current.state() == next.state()
-                    && current.stage() == next.stage()
-                    && Objects.equals(current.stageMessage(), next.stageMessage())
-                    && current.completedUnits() == next.completedUnits()
-                    && current.totalUnits() == next.totalUnits()
-                    && Objects.equals(current.currentSourcePath(), next.currentSourcePath());
-        }
-    }
 }
